[IMP] core: make author field mandatory in manifests
External contributors (and namely Odoo PS) forget to set the author
field in their manifest file. The result is that the module is wrongly
registered with `'author': 'Odoo S.A.'`.

This work makes the field mandatory, with a warning when not set. It
also offers an upgrade-code script to mass-rewrite all the manifests in
order to set the author.

task-4485983